[US-197] correcting css
