+++ frontend/src/app/components/hand-history/HandHistory.module.scss
@@ -1,30 +1,32 @@
 @import '../../assets/styles/variables.scss';
 
 .header,
-.playerList {
+.historyList {
   display: flex;
-  border-bottom-radius: 5px;
-  padding-left: 30px;
-  height: 400px;
-  max-height: 400px;
-  border-color: #f7f0e1;
+  border: 1px solid $white-color;
 }
 
 .header {
+  font-size: $font-size-header;
+  font-weight: bold;
   height: 60px;
   align-items: center;
   justify-content: center;
-  margin-bottom: 2px;
-  border-top-radius: 5px;
-  border-color: #f7f0e1;
+  border-top-left-radius: 5px;
+  border-top-right-radius: 5px;
 }
 
 .historyList {
+  padding-left: 30px;
+  padding-block: 30px;
   flex-direction: column;
   overflow-y: auto;
   overflow-x: hidden;
-  padding-right: 30px;
+  height: 50vh;
   overflow: auto;
+  padding-right: 30px;
+  border-bottom-left-radius: 5px;
+  border-bottom-right-radius: 5px;
 }
 
 .historyItem {
