+++ frontend/src/app/components/information-container/InformationContainer.module.scss
@@ -9,7 +9,5 @@
 
 .container > * {
   background-color: $transparent-white-color;
-  padding-left: 30px;
-  padding-block: 30px;
   color: $blue-color;
 }
