+++ frontend/src/app/components/information-panel/InformationPanel.tsx
@@ -4,6 +4,8 @@ import NextLevelInfos from '../next-level-infos/NextLevelInfos'
 import PlayersStatus from '../players-status/PlayersStatus'
 import { LevelInfosModel } from '@/app/models/LevelInfos.model'
 import { PlayerHandInfosModel } from '@/app/models/PlayerHandInfos.model'
+import HandHistory from '@/app/components/hand-history/HandHistory'
+import { HandHistoryModel } from '@/app/models/HandHistory.model'
 import { GameActionEnum } from '@/app/enums/GameAction.enum'
 
 export default function InformationPanel() {
@@ -34,10 +36,89 @@ export default function InformationPanel() {
       betIsValid: true,
     },
   ]
+  const mockHandHistory: HandHistoryModel[] = [
+    {
+      seat: 3,
+      action: GameActionEnum.CHECK,
+      betValue: 0,
+    },
+    {
+      seat: 4,
+      action: GameActionEnum.BET,
+      betValue: 300,
+    },
+    {
+      seat: 6,
+      action: GameActionEnum.CALL,
+      betValue: 300,
+    },
+    {
+      seat: 1,
+      action: GameActionEnum.FOLD,
+      betValue: 0,
+    },
+    {
+      seat: 2,
+      action: GameActionEnum.CALL,
+      betValue: 300,
+    },
+    {
+      seat: 3,
+      action: GameActionEnum.CHECK,
+      betValue: 0,
+    },
+    {
+      seat: 4,
+      action: GameActionEnum.BET,
+      betValue: 300,
+    },
+    {
+      seat: 6,
+      action: GameActionEnum.CALL,
+      betValue: 300,
+    },
+    {
+      seat: 1,
+      action: GameActionEnum.FOLD,
+      betValue: 0,
+    },
+    {
+      seat: 2,
+      action: GameActionEnum.CALL,
+      betValue: 300,
+    },
+    {
+      seat: 3,
+      action: GameActionEnum.CHECK,
+      betValue: 0,
+    },
+    {
+      seat: 4,
+      action: GameActionEnum.BET,
+      betValue: 300,
+    },
+    {
+      seat: 6,
+      action: GameActionEnum.CALL,
+      betValue: 300,
+    },
+    {
+      seat: 1,
+      action: GameActionEnum.FOLD,
+      betValue: 0,
+    },
+    {
+      seat: 2,
+      action: GameActionEnum.CALL,
+      betValue: 300,
+    },
+  ]
 
   return (
     <Box className={styles.informationPanel}>
-      <Box className={styles.leftInformationPanel}></Box>
+      <Box className={styles.leftInformationPanel}>
+        <HandHistory HandHistoryInfos={mockHandHistory} />
+      </Box>
       <Box className={styles.middleInformationPanel}>
         <NextLevelInfos levelInfos={mockLevelInfos} />
       </Box>
